infrastructure/requirements.txt
aws-cdk-lib>=2.80.0
constructs>=10.2.0
aws-cdk.aws-apigateway>=2.80.0
aws-cdk.aws-lambda>=2.80.0
aws-cdk.aws-dynamodb>=2.80.0
aws-cdk.aws-kinesis>=2.80.0
aws-cdk.aws-eks>=2.80.0
aws-cdk.aws-cognito>=2.80.0
aws-cdk.aws-s3>=2.80.0
aws-cdk.aws-iam>=2.80.0
aws-cdk.aws-logs>=2.80.0
aws-cdk.aws-ec2>=2.80.0
aws-cdk.aws-cloudwatch>=2.80.0 